Refactor - moved HumanWizardToken to shared area
